Moved the telemetry plugin package
