+++ test/framework/src/main/java/org/opensearch/test/telemetry/MockTelemetryPlugin.java
@@ -6,11 +6,13 @@
  * compatible open source license.
  */
 
-package org.opensearch.telemetry;
+package org.opensearch.test.telemetry;
 
 import java.util.Optional;
 import org.opensearch.plugins.Plugin;
 import org.opensearch.plugins.TelemetryPlugin;
+import org.opensearch.telemetry.Telemetry;
+import org.opensearch.telemetry.TelemetrySettings;
 
 /**
  * Mock {@link TelemetryPlugin} implementation for testing.
